requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
art==5.7
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.3
iso8601==1.0.2
multidict==6.0.2
py-cord==2.1.3
pycord-multicog==1.0.2
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.2.1
time-str==0.1.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
yarl==1.8.1
